Reader menu changes & quick fixes
This commit adds a new item to the reader's menu, "返回主菜单". It has the same effect as the previous "退出"; and the "退出" now brings the calculator back to home menu.
It also fixes a bug that when the reader returned prematurely, the menu-key-to-return flag isn't toggled back on and user cannot go back to their home menu.