New bookmark structure
The bookmark structure now records both the page location and the byte location, so when the font size is changed, your bookmarks will not be cleared, but be converted instead. Note that the bookmark is still page-based, not line-based, so when you change your font, the previous bookmark won't precisely lead you to the place you want.